+++ ank/__init__.py
@@ -1,23 +1,23 @@
 __author__ = 'sunary'
 
 
-VERSION = '1.6.0'
+VERSION = '2.0.0'
 API_DEFAULT_PORT = '5372'
 
 
-from components.pipe_app import PipeApp
-from components.api_app import APIApp
-from components.schedule_app import ScheduleApp
+from ank.components.pipe_app import PipeApp
+from ank.components.api_app import APIApp
+from ank.components.schedule_app import ScheduleApp
 
-from components.join_app import JoinApp
-from components.split_app import SplitApp
-from components.log_app import LogApp
+from ank.components.join_app import JoinApp
+from ank.components.split_app import SplitApp
+from ank.components.log_app import LogApp
 
-from components.redis_subscribe import RedisSubscribe
-from components.kafka_consumer import KafkaAnkConsumer
-from components.rabbitmq_consumer import RabbitmqConsumer
-from components.zmq_consumer import ZeroMqConsumer
+from ank.components.redis_subscribe import RedisSubscribe
+from ank.components.kafka_consumer import KafkaAnkConsumer
+from ank.components.rabbitmq_consumer import RabbitmqConsumer
+from ank.components.zmq_consumer import ZeroMqConsumer
 
-from components.kafka_producer import KafkaAnkProducer
-from components.rabbitmq_producer import RabbitmqProducer
-from components.zmq_producer import ZeroMqProducer
+from ank.components.kafka_producer import KafkaAnkProducer
+from ank.components.rabbitmq_producer import RabbitmqProducer
+from ank.components.zmq_producer import ZeroMqProducer