+++ ank/tests/processor.py
@@ -56,7 +56,7 @@ class ConditionalApp(PipeApp):
 
 class OddApp(PipeApp):
 
-    def init_app(self, *agrs):
+    def init_app(self, *args):
         pass
 
     def run(self, process=None):