+++ examples/api_app/processor.py
@@ -7,8 +7,8 @@ from ank.components.pipe_app import PipeApp
 
 class ExampleApp(PipeApp):
 
-    def __init__(self, agrs, **kwagrs):
-        super(ExampleApp, self).__init__()
+    def init_app(self, mongo_client=None):
+        self.mongo = mongo_client
 
     def start(self):
         api_app = ExampleAPI(host='localhost', port=5372, mongo_client=self.mongo, mongo_db='demo')